test: refactor rendered audio assertions
